refactor: remove old limit order review modal and extract recap row
